requirements.txt
## ATLAS: Python Dependencies
## AMD Developer Hackathon 2026
## Python 3.12+

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# AI/ML Orchestration
crewai==0.1.0
langchain==0.0.340
langchain-community==0.0.10
langchain-ollama>=0.1.0

# LLM Providers
ollama==0.1.0
openai>=1.52.0
httpx>=0.25.0,<0.28.0

# Document Processing
pypdf==3.17.1
pillow==10.1.0
PyMuPDF>=1.23.0
pymupdf4llm>=0.0.5
pytesseract>=0.3.10
pdf2image>=1.16.3

# Database
supabase==2.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Embeddings & Search
sentence-transformers==2.2.2
numpy==1.24.3
scipy==1.11.4

# Web Framework
streamlit==1.28.1
python-multipart==0.0.9

# Utilities
requests==2.31.0
aiohttp==3.9.1
pydantic-settings==2.1.0

# Logging
python-json-logger==2.0.7
loguru>=0.7.0

# Dev/Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
